AArch64 (arm64) support for image building, mock
- mock:
  * extend config_opts['legal_host_arches'] to include 'aarch64';
  * parameterize RPM target architecture based on the output of
    $(uname -m) in create_mock_config.sh;
- Dockerfile-dib:
  * Bump base to centos:7.6.1810, but since centos:7.x.y tags are
    broken on arm64 [1], apply an arch-specific patch to the DIB
    Dockerfile which pins the FROM image to a specific sha256;
  * Use AArch64 altsig repository for openstack-queens packages;
  * Bump diskimage-builder to 2.26.1 to include arm64 critical fixes;
  * Add tools required for UEFI partition handling and other AArch64
    specific requirements to the arm64 container image (gdisk,
    dosfstools, util-linux, qemu-img);
- dib elements:
  * finalise.d/99-collect-rpm-info: AArch64 workaround for packages
    initially installed from 'updates' repo residing in 'base' repo
    after all distro/repo handling take place;
- create_golden_image.sh:
  * dib: Add additional element block-device-efi on AArch64 (only);
- build_step_create_install_cd.sh:
  * parameterize CentOS artifacts URI based on machine architecture;
  * skip isolinux artifacts download/handling for AArch64, since
    PC BIOS is x86 specific;
  * ISO(s) build: AArch64 only supports UEFI booting, skip isolinux;
  * AArch64: use legacy method of copying out the contents of "/boot"
    from the golden image since the AArch64 image has a more complex
    partition layout and would require mounting both "/boot" and ESP;
- create_rpmdata_in_docker.sh:
  * fix bash shebang since script uses bash-isms;
- tools/buildconfig.py:
  * add support for replacing the magic string '#ARCH#' with current
    platform architecture in option names;

+++ build_step_create_install_cd.sh
@@ -30,9 +30,10 @@ mkdir -p $tmp
 rm -rf $iso_build_dir
 mkdir -p $iso_build_dir
 
+iso_arch="$(uname -m)"
 reposnap_base=$(_read_build_config DEFAULT centos_reposnap)
 release_version=$PRODUCT_RELEASE_LABEL
-reposnap_base_dir="${reposnap_base}/os/x86_64/"
+reposnap_base_dir="${reposnap_base}/os/${iso_arch}/"
 iso_image_label=$(_read_build_config DEFAULT iso_image_label)
 cd_efi_dir="${reposnap_base_dir}/EFI"
 cd_images_dir="${reposnap_base_dir}/images"
@@ -59,48 +60,63 @@ wget_dir ${cd_efi_dir}/
 wget_dir ${cd_images_dir}/
 rm -f images/boot.iso
 sync
-wget_dir ${cd_isolinux_dir}/
-chmod +w -R isolinux/ EFI/ images/
+chmod +w -R EFI/ images/
+# AArch64 does not support PC-BIOS, so skip all isolinux processing
+if [ "${iso_arch}" != 'aarch64' ]; then
+    wget_dir ${cd_isolinux_dir}/
+    chmod +w -R isolinux/
 
-if [ -e $scriptdir/isolinux/isolinux.cfg ]; then
-    cp $scriptdir/isolinux/isolinux.cfg isolinux/isolinux.cfg
-else
-    sed -i "s/^timeout.*/timeout 100/" isolinux/isolinux.cfg
-    sed -i "s/^ -  Press.*/Beginning the cloud installation process/" isolinux/boot.msg
-    sed -i "s/^#menu hidden/menu hidden/" isolinux/isolinux.cfg
-    sed -i "s/menu default//" isolinux/isolinux.cfg
-    sed -i "/^label linux/amenu default" isolinux/isolinux.cfg
-    sed -i "/append initrd/ s/$/ console=tty0 console=ttyS1,115200/" isolinux/isolinux.cfg
+    if [ -e $scriptdir/isolinux/isolinux.cfg ]; then
+        cp $scriptdir/isolinux/isolinux.cfg isolinux/isolinux.cfg
+    else
+        sed -i "s/^timeout.*/timeout 100/" isolinux/isolinux.cfg
+        sed -i "s/^ -  Press.*/Beginning the cloud installation process/" isolinux/boot.msg
+        sed -i "s/^#menu hidden/menu hidden/" isolinux/isolinux.cfg
+        sed -i "s/menu default//" isolinux/isolinux.cfg
+        sed -i "/^label linux/amenu default" isolinux/isolinux.cfg
+        sed -i "/append initrd/ s/$/ console=tty0 console=ttyS1,115200/" isolinux/isolinux.cfg
+    fi
+    cp -f $scriptdir/akraino_splash.png isolinux/splash.png
 fi
-cp -f $scriptdir/akraino_splash.png isolinux/splash.png
 
 popd
 
 pushd $tmp
 
  # Copy latest kernel and initrd-provisioning from boot dir
-qemu-img convert $input_image guest-image.raw
-myloop=$(sudo losetup -fP --show guest-image.raw)
-mkdir mnt
-sudo mount -o loop ${myloop}p1 mnt/
-sudo rsync -avA mnt/boot .
-sudo chown -R $(id -u):$(id -g) boot
-sudo umount mnt
-sudo losetup -d ${myloop}
-rm -f guest-image.raw
+if [ "${iso_arch}" != 'aarch64' ]; then
+    qemu-img convert $input_image guest-image.raw
+    myloop=$(sudo losetup -fP --show guest-image.raw)
+    mkdir mnt
+    sudo mount -o loop ${myloop}p1 mnt/
+    sudo rsync -avA mnt/boot .
+    sudo chown -R $(id -u):$(id -g) boot
+    sudo umount mnt
+    sudo losetup -d ${myloop}
+    rm -f guest-image.raw
+else
+    export LIBGUESTFS_BACKEND=direct
+    virt-copy-out -a $input_image /boot/ ./
+fi
 
 chmod u+w boot/
-rm -f $iso_build_dir/isolinux/vmlinuz $iso_build_dir/isolinux/initrd.img
 KVER=`ls -lrt boot/vmlinuz-* |grep -v rescue |tail -n1 |awk -F 'boot/vmlinuz-' '{print $2}'`
-cp -fp boot/vmlinuz-${KVER} $iso_build_dir/isolinux/vmlinuz
-cp -fp boot/initrd-provisioning.img $iso_build_dir/isolinux/initrd.img
+if [ "${iso_arch}" != 'aarch64' ]; then
+    rm -f $iso_build_dir/isolinux/vmlinuz $iso_build_dir/isolinux/initrd.img
+    cp -fp boot/vmlinuz-${KVER} $iso_build_dir/isolinux/vmlinuz
+    cp -fp boot/initrd-provisioning.img $iso_build_dir/isolinux/initrd.img
+fi
 rm -rf boot/
 
 echo "Generating boot iso"
+if [ "${iso_arch}" != 'aarch64' ]; then
+    bios_specific_args="-b isolinux/isolinux.bin -c isolinux/boot.cat \
+                        -no-emul-boot -boot-load-size 4 -boot-info-table"
+fi
 genisoimage  -U -r -v -T -J -joliet-long \
   -V "${release_version}" -A "${release_version}" -P ${iso_image_label} \
-  -b isolinux/isolinux.bin -c isolinux/boot.cat -no-emul-boot -boot-load-size 4 \
-  -boot-info-table -eltorito-alt-boot -e images/efiboot.img -no-emul-boot \
+  ${bios_specific_args:-} \
+  -eltorito-alt-boot -e images/efiboot.img -no-emul-boot \
   -o boot.iso $iso_build_dir
 _publish_image $tmp/boot.iso $output_bootcd_path
 
@@ -112,10 +128,10 @@ mkdir -p $iso_build_dir/rpms
 echo "Generating product iso"
 genisoimage  -U -r -v -T -J -joliet-long \
   -V "${release_version}" -A "${release_version}" -P ${iso_image_label} \
-  -b isolinux/isolinux.bin -c isolinux/boot.cat -no-emul-boot -boot-load-size 4 \
-  -boot-info-table -eltorito-alt-boot -e images/efiboot.img -no-emul-boot \
+  ${bios_specific_args:-} \
+  -eltorito-alt-boot -e images/efiboot.img -no-emul-boot \
   -o release.iso $iso_build_dir
-isohybrid $tmp/release.iso
+[ "${iso_arch}" = 'aarch64' ] || isohybrid $tmp/release.iso
 _publish_image $tmp/release.iso $output_image_path
 
 echo "Clean up to preserve workspace footprint"
